LayoutTests/fast/selectors/attribute-list-value-matching-is-ascii-case-insensitive-expected.txt

When matching attributes case insensitively, it should be ASCII case insensitive. This test verifies the behavior when matching a substring in a whitespace-separated list (e.g. [a~="b"])

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing "[data-attribute~=WÈb]"
PASS document.querySelectorAll('#test-root [data-attribute~=WÈb]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute~=WÈb]')[0].id is "target1"
PASS styledElements is ["target1"]

Testing "[data-attribute~=Kit]"
PASS document.querySelectorAll('#test-root [data-attribute~=Kit]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute~=Kit]')[0].id is "target1"
PASS styledElements is ["target1"]

Testing "[data-attribute~=wÈb]"
PASS document.querySelectorAll('#test-root [data-attribute~=wÈb]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute~=wÈb]')[0].id is "target2"
PASS styledElements is ["target2"]

Testing "[data-attribute~=kit]"
PASS document.querySelectorAll('#test-root [data-attribute~=kit]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute~=kit]')[0].id is "target2"
PASS styledElements is ["target2"]

Testing "[data-attribute~=WÈB]"
PASS document.querySelectorAll('#test-root [data-attribute~=WÈB]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute~=WÈB]')[0].id is "target3"
PASS styledElements is ["target3"]

Testing "[data-attribute~=KIT]"
PASS document.querySelectorAll('#test-root [data-attribute~=KIT]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute~=KIT]')[0].id is "target3"
PASS styledElements is ["target3"]

Testing "[data-attribute~=Wèb]"
PASS document.querySelectorAll('#test-root [data-attribute~=Wèb]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute~=Wèb]')[0].id is "target4"
PASS styledElements is ["target4"]

Testing "[data-attribute~=wèb]"
PASS document.querySelectorAll('#test-root [data-attribute~=wèb]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute~=wèb]')[0].id is "target5"
PASS styledElements is ["target5"]

Testing "[data-attribute~=kIt]"
PASS document.querySelectorAll('#test-root [data-attribute~=kIt]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute~=kIt]')[0].id is "target5"
PASS styledElements is ["target5"]

Testing "[data-attribute~=WÈb i]"
PASS document.querySelectorAll('#test-root [data-attribute~=WÈb i]').length is 3
PASS document.querySelectorAll('#test-root [data-attribute~=WÈb i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute~=WÈb i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute~=WÈb i]')[2].id is "target3"
PASS styledElements is ["target1","target2","target3"]

Testing "[data-attribute~=Kit i]"
PASS document.querySelectorAll('#test-root [data-attribute~=Kit i]').length is 4
PASS document.querySelectorAll('#test-root [data-attribute~=Kit i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute~=Kit i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute~=Kit i]')[2].id is "target3"
PASS document.querySelectorAll('#test-root [data-attribute~=Kit i]')[3].id is "target5"
PASS styledElements is ["target1","target2","target3","target5"]

Testing "[data-attribute~=wÈb i]"
PASS document.querySelectorAll('#test-root [data-attribute~=wÈb i]').length is 3
PASS document.querySelectorAll('#test-root [data-attribute~=wÈb i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute~=wÈb i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute~=wÈb i]')[2].id is "target3"
PASS styledElements is ["target1","target2","target3"]

Testing "[data-attribute~=kit i]"
PASS document.querySelectorAll('#test-root [data-attribute~=kit i]').length is 4
PASS document.querySelectorAll('#test-root [data-attribute~=kit i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute~=kit i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute~=kit i]')[2].id is "target3"
PASS document.querySelectorAll('#test-root [data-attribute~=kit i]')[3].id is "target5"
PASS styledElements is ["target1","target2","target3","target5"]

Testing "[data-attribute~=WÈB i]"
PASS document.querySelectorAll('#test-root [data-attribute~=WÈB i]').length is 3
PASS document.querySelectorAll('#test-root [data-attribute~=WÈB i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute~=WÈB i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute~=WÈB i]')[2].id is "target3"
PASS styledElements is ["target1","target2","target3"]

Testing "[data-attribute~=KIT i]"
PASS document.querySelectorAll('#test-root [data-attribute~=KIT i]').length is 4
PASS document.querySelectorAll('#test-root [data-attribute~=KIT i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute~=KIT i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute~=KIT i]')[2].id is "target3"
PASS document.querySelectorAll('#test-root [data-attribute~=KIT i]')[3].id is "target5"
PASS styledElements is ["target1","target2","target3","target5"]

Testing "[data-attribute~=Wèb i]"
PASS document.querySelectorAll('#test-root [data-attribute~=Wèb i]').length is 2
PASS document.querySelectorAll('#test-root [data-attribute~=Wèb i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [data-attribute~=Wèb i]')[1].id is "target5"
PASS styledElements is ["target4","target5"]

Testing "[data-attribute~=wèb i]"
PASS document.querySelectorAll('#test-root [data-attribute~=wèb i]').length is 2
PASS document.querySelectorAll('#test-root [data-attribute~=wèb i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [data-attribute~=wèb i]')[1].id is "target5"
PASS styledElements is ["target4","target5"]

Testing "[data-attribute~=kIt i]"
PASS document.querySelectorAll('#test-root [data-attribute~=kIt i]').length is 4
PASS document.querySelectorAll('#test-root [data-attribute~=kIt i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute~=kIt i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute~=kIt i]')[2].id is "target3"
PASS document.querySelectorAll('#test-root [data-attribute~=kIt i]')[3].id is "target5"
PASS styledElements is ["target1","target2","target3","target5"]

Testing "[multiple~=WèB]"
PASS document.querySelectorAll('#test-root [multiple~=WèB]').length is 3
PASS document.querySelectorAll('#test-root [multiple~=WèB]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple~=WèB]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple~=WèB]')[2].id is "target3"
PASS styledElements is ["target1","target2","target3"]

Testing "[multiple~=KIT]"
PASS document.querySelectorAll('#test-root [multiple~=KIT]').length is 4
PASS document.querySelectorAll('#test-root [multiple~=KIT]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple~=KIT]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple~=KIT]')[2].id is "target3"
PASS document.querySelectorAll('#test-root [multiple~=KIT]')[3].id is "target4"
PASS styledElements is ["target1","target2","target3","target4"]

Testing "[multiple~=Wèb]"
PASS document.querySelectorAll('#test-root [multiple~=Wèb]').length is 3
PASS document.querySelectorAll('#test-root [multiple~=Wèb]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple~=Wèb]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple~=Wèb]')[2].id is "target3"
PASS styledElements is ["target1","target2","target3"]

Testing "[multiple~=Kit]"
PASS document.querySelectorAll('#test-root [multiple~=Kit]').length is 4
PASS document.querySelectorAll('#test-root [multiple~=Kit]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple~=Kit]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple~=Kit]')[2].id is "target3"
PASS document.querySelectorAll('#test-root [multiple~=Kit]')[3].id is "target4"
PASS styledElements is ["target1","target2","target3","target4"]

Testing "[multiple~=Wèb]"
PASS document.querySelectorAll('#test-root [multiple~=Wèb]').length is 3
PASS document.querySelectorAll('#test-root [multiple~=Wèb]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple~=Wèb]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple~=Wèb]')[2].id is "target3"
PASS styledElements is ["target1","target2","target3"]

Testing "[multiple~=kit]"
PASS document.querySelectorAll('#test-root [multiple~=kit]').length is 4
PASS document.querySelectorAll('#test-root [multiple~=kit]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple~=kit]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple~=kit]')[2].id is "target3"
PASS document.querySelectorAll('#test-root [multiple~=kit]')[3].id is "target4"
PASS styledElements is ["target1","target2","target3","target4"]

Testing "[multiple~=wÈb]"
PASS document.querySelectorAll('#test-root [multiple~=wÈb]').length is 2
PASS document.querySelectorAll('#test-root [multiple~=wÈb]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple~=wÈb]')[1].id is "target5"
PASS styledElements is ["target4","target5"]

Testing "[multiple~=kiT]"
PASS document.querySelectorAll('#test-root [multiple~=kiT]').length is 4
PASS document.querySelectorAll('#test-root [multiple~=kiT]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple~=kiT]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple~=kiT]')[2].id is "target3"
PASS document.querySelectorAll('#test-root [multiple~=kiT]')[3].id is "target4"
PASS styledElements is ["target1","target2","target3","target4"]

Testing "[multiple~=wÈb]"
PASS document.querySelectorAll('#test-root [multiple~=wÈb]').length is 2
PASS document.querySelectorAll('#test-root [multiple~=wÈb]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple~=wÈb]')[1].id is "target5"
PASS styledElements is ["target4","target5"]

Testing "[multiple~=WèB i]"
PASS document.querySelectorAll('#test-root [multiple~=WèB i]').length is 3
PASS document.querySelectorAll('#test-root [multiple~=WèB i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple~=WèB i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple~=WèB i]')[2].id is "target3"
PASS styledElements is ["target1","target2","target3"]

Testing "[multiple~=KIT i]"
PASS document.querySelectorAll('#test-root [multiple~=KIT i]').length is 4
PASS document.querySelectorAll('#test-root [multiple~=KIT i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple~=KIT i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple~=KIT i]')[2].id is "target3"
PASS document.querySelectorAll('#test-root [multiple~=KIT i]')[3].id is "target4"
PASS styledElements is ["target1","target2","target3","target4"]

Testing "[multiple~=Wèb i]"
PASS document.querySelectorAll('#test-root [multiple~=Wèb i]').length is 3
PASS document.querySelectorAll('#test-root [multiple~=Wèb i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple~=Wèb i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple~=Wèb i]')[2].id is "target3"
PASS styledElements is ["target1","target2","target3"]

Testing "[multiple~=Kit i]"
PASS document.querySelectorAll('#test-root [multiple~=Kit i]').length is 4
PASS document.querySelectorAll('#test-root [multiple~=Kit i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple~=Kit i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple~=Kit i]')[2].id is "target3"
PASS document.querySelectorAll('#test-root [multiple~=Kit i]')[3].id is "target4"
PASS styledElements is ["target1","target2","target3","target4"]

Testing "[multiple~=Wèb i]"
PASS document.querySelectorAll('#test-root [multiple~=Wèb i]').length is 3
PASS document.querySelectorAll('#test-root [multiple~=Wèb i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple~=Wèb i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple~=Wèb i]')[2].id is "target3"
PASS styledElements is ["target1","target2","target3"]

Testing "[multiple~=kit i]"
PASS document.querySelectorAll('#test-root [multiple~=kit i]').length is 4
PASS document.querySelectorAll('#test-root [multiple~=kit i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple~=kit i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple~=kit i]')[2].id is "target3"
PASS document.querySelectorAll('#test-root [multiple~=kit i]')[3].id is "target4"
PASS styledElements is ["target1","target2","target3","target4"]

Testing "[multiple~=wÈb i]"
PASS document.querySelectorAll('#test-root [multiple~=wÈb i]').length is 2
PASS document.querySelectorAll('#test-root [multiple~=wÈb i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple~=wÈb i]')[1].id is "target5"
PASS styledElements is ["target4","target5"]

Testing "[multiple~=kiT i]"
PASS document.querySelectorAll('#test-root [multiple~=kiT i]').length is 4
PASS document.querySelectorAll('#test-root [multiple~=kiT i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple~=kiT i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple~=kiT i]')[2].id is "target3"
PASS document.querySelectorAll('#test-root [multiple~=kiT i]')[3].id is "target4"
PASS styledElements is ["target1","target2","target3","target4"]

Testing "[multiple~=wÈb i]"
PASS document.querySelectorAll('#test-root [multiple~=wÈb i]').length is 2
PASS document.querySelectorAll('#test-root [multiple~=wÈb i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple~=wÈb i]')[1].id is "target5"
PASS styledElements is ["target4","target5"]

PASS successfullyParsed is true

TEST COMPLETE
